requirements.txt
selenium==4.4.3
paddleocr~=2.6.0.1
paddlepaddle==2.3.2
chromedriver-autoinstaller==0.4.0
python-slugify==6.1.2
